Config file revamp
I found that the original had simultaneously too much customization and not enough. Lost of file path options but very limited ways of interacting with the components that make up the workflow.

My philosophy here is that it's better to completely expose the settings for the libraries being used than to offer only a few specific properties—less to maintain when options are added or changed.

I'm still not sold on the utility of allowing fundamental file path changes; the goal of this scaffolding is rapid development and some amount of uniformity. The one use case that stands out in my mind is the possibility of applying an update to a legacy project and the default folder structure changing, but I'm not sure backwards-compatibility is a goal. I'm also not sold on vendor script concatenation, but I'm leaving the options for now. I have removed the "custom" JS path/filename options because a single compiled JS file is extremely limiting and inefficient (which lead me to tinkering with this in the first place).

Some variable names changed either for consistency or because the semantics were off. For instance, `projectURL` became `localDomain` because this is not a URL, it's a domain, specifically a local one. Also, why was a single variable all caps? These are all "constants".

Comments have been reformatted into a standard docblock-like structure for readability.

I've also conformed this to the modern ES6 export format.